# openapi/fusion-integrated.v1.yaml
openapi: 3.0.3
info:
  title: Fusion-Integrated API
  version: 1.0.0
  description: Non-academic platform. Independent modules, one login.
paths:
...
  /api/v1/placement/students/cpi:
    get:
      operationId: v1_placement_students_cpi_retrieve
      description: GET /api/v1/placement/students/cpi
      tags:
      - v1
      security:
      - iamSession: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcademicDirectory'
          description: ''
...
components:
  schemas:
    AcademicDirectory:
      type: object
      properties:
        count:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        results:
          type: array
          items:
            $ref: '#/components/schemas/AcademicStandingRow'
      required:
      - count
      - limit
      - offset
      - results
...
    AcademicStandingRow:
      type: object
      description: |-
        One row of the CPI directory. `cpi` is null when nothing is declared —
        never zero.
      properties:
        user_id:
          type: integer
        roll_no:
          type: string
        name:
          type: string
        programme:
          type: string
        discipline:
          type: string
        batch_year:
          type: integer
          nullable: true
        cpi:
          type: string
          nullable: true
        semester:
          type: integer
          nullable: true
        semester_type:
          type: string
          nullable: true
        declared_seq:
          type: integer
          nullable: true
        earned_credits:
          type: string
          nullable: true
        active_backlogs:
          type: integer
          nullable: true
      required:
      - active_backlogs
      - batch_year
      - cpi
      - declared_seq
      - discipline
      - earned_credits
      - name
      - programme
      - roll_no
      - semester
      - semester_type
      - user_id
...
  securitySchemes:
    iamSession:
      type: apiKey
      in: header
      name: Authorization
      description: '`Token <value>` from the IAM, or the httpOnly `auth_token` cookie
        the SPA carries.'
    recruiterSession:
      type: apiKey
      in: header
      name: Authorization
      description: '`Recruiter <value>` for an external company account, or the httpOnly
        `recruiter_session` cookie set at portal login. A separate pool from institute
        users.'